Add util class SortedCollection and related tests
A SortedCollection is a collection of elements that allows duplicates and nulls and will always be in sorted order, regardless of how elements are added to it. It implements the Collection interface. The type of elements belonging to a SortedCollection must implement the Comparable interface so that insertion sorting can occur.